src/app/my-orders/my-orders.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { RouterTestingModule } from '@angular/router/testing';
import { MyOrdersComponent } from './my-orders.component';
import { AuthService } from '../auth.service';

describe('MyOrdersComponent', () => {
  let component: MyOrdersComponent;
  let fixture: ComponentFixture<MyOrdersComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [
        HttpClientTestingModule,
        RouterTestingModule,
        MyOrdersComponent
      ],
      providers: [AuthService]
    }).compileComponents();

    fixture = TestBed.createComponent(MyOrdersComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  // 1. Check if component is created successfully
  it('should create the component', () => {
    expect(component).toBeTruthy();
  });

  // 2. Check initial state of properties
  it('should initialize with default values', () => {
    expect(component.orders).toEqual([]); // Adjust based on your actual component's default state
  });

  // 3. Check if the component has a defined property (e.g., `title`)
  it('should have a defined title', () => {
    // Replace `title` with an actual property from your component
    expect(component.title).toBeDefined();
  });

  // 4. Check if a method can be called successfully
});
